ICovidAPIv1.py
For the Factory Design Pattern
